+++ filebeat/input/filestream/filestream_test.go
@@ -110,77 +110,6 @@ func TestLogFileTruncated(t *testing.T) {
 	assert.Equal(t, ErrFileTruncate, err)
 }
 
-func TestLogFileRenamed(t *testing.T) {
-	f := createTestLogFile()
-	defer f.Close()
-
-	renamedFile := f.Name() + ".renamed"
-
-	reader, err := newFileReader(
-		logp.L(),
-		context.TODO(),
-		f,
-		readerConfig{},
-		closerConfig{
-			OnStateChange: stateChangeCloserConfig{
-				CheckInterval: 1 * time.Second,
-				Renamed:       true,
-			},
-		},
-	)
-	if err != nil {
-		t.Fatalf("error while creating logReader: %+v", err)
-	}
-
-	buf := make([]byte, 1024)
-	_, err = reader.Read(buf)
-	assert.Nil(t, err)
-
-	err = os.Rename(f.Name(), renamedFile)
-	if err != nil {
-		t.Fatalf("error while renaming file: %+v", err)
-	}
-
-	err = readUntilError(reader)
-	os.Remove(renamedFile)
-
-	assert.Equal(t, ErrClosed, err)
-}
-
-func TestLogFileRemoved(t *testing.T) {
-	f := createTestLogFile()
-	defer f.Close()
-
-	reader, err := newFileReader(
-		logp.L(),
-		context.TODO(),
-		f,
-		readerConfig{},
-		closerConfig{
-			OnStateChange: stateChangeCloserConfig{
-				CheckInterval: 1 * time.Second,
-				Removed:       true,
-			},
-		},
-	)
-	if err != nil {
-		t.Fatalf("error while creating logReader: %+v", err)
-	}
-
-	buf := make([]byte, 1024)
-	_, err = reader.Read(buf)
-	assert.Nil(t, err)
-
-	err = os.Remove(f.Name())
-	if err != nil {
-		t.Fatalf("error while remove file: %+v", err)
-	}
-
-	err = readUntilError(reader)
-
-	assert.Equal(t, ErrClosed, err)
-}
-
 func createTestLogFile() *os.File {
 	f, err := ioutil.TempFile("", "filestream_reader_test")
 	if err != nil {
